walk packfile support

+++ grawler.sh
@@ -12,19 +12,36 @@ PW_EXTRACT='-i password'
 SECRET_EXTRACT='-i secret'
 KEY_EXTRACT='-i key'
 COMMITS=false
+OBJECT_HASH=
+WALK_PACK=false
+RESUME=false
 
 SCRIPT_DIR=`pwd -P`
 
 
 usage() {
-	echo "usage: $program_name [-shC] [-g dir] [-w dir] [-f filter] [-x regex]"
+	echo "usage: $program_name [-shCPr] [-g dir] [-w dir] [-f filter] [-x regex] [-W hash]"
 	echo "	-g 	git directory"
 	echo "	-w 	working directory"
 	echo "	-f 	filter for git log"
 	echo "	-x 	extract: (p) Password, (k) Keys, (c) Secrets, (s) SSN"
 	echo "	-h 	print this cruft"
 	echo "  -C 	print commit hashes"
+	echo "	-W 	which commit has hash object"
+	echo " 	-P 	walk pack file"
+	echo " 	-r 	resume (don't kill tree_file"
 	echo "Only one type of extract may be performed at a time"
+}
+
+which_commit() {
+	obj_name="$1"
+	shift
+	git log "$@" --pretty=format:'%T %H %s' \
+	| while read tree commit subject ; do
+	    if git ls-tree -r $tree | grep -q "$obj_name" ; then
+	        echo $tree $commit "$subject"
+	    fi
+	done
 }
 
 dump_blob() {
@@ -33,7 +50,7 @@ dump_blob() {
 	# but to make things more flexible started using python extractor.py instead of awk
 	# hopefully we can condense this then
 	commit_hash=$1
-	if [ $EXTRACT == "s" ]; then
+	if [ "$EXTRACT" == "s" ]; then
 		if [ "$COMMITS" = true ]; then
 			git cat-file -p $1 | egrep '[0-9]{3}-[0-9]{2}-[0-9]{4}' | python ${SCRIPT_DIR}/extractor.py --ssn -H $commit_hash
 		else
@@ -41,11 +58,11 @@ dump_blob() {
 			git cat-file -p $1 | egrep '[0-9]{3}-[0-9]{2}-[0-9]{4}' | python ${SCRIPT_DIR}/extractor.py --ssn
 				# awk 'match($0, /[0-9]{3}-[0-9]{2}-[0-9]{4}/) { print substr( $0, RSTART, RLENGTH)}'
 		fi
-	elif [ $EXTRACT == "p" ]; then
+	elif [ "$EXTRACT" == "p" ]; then
 		git cat-file -p $1 | egrep -i 'password|pw' | python ${SCRIPT_DIR}/extractor.py --password
-	elif [ $EXTRACT == "k" ]; then
+	elif [ "$EXTRACT" == "k" ]; then
 		git cat-file -p $1 | egrep -i 'key' | python ${SCRIPT_DIR}/extractor.py --key
-	elif [ $EXTRACT == "c" ]; then
+	elif [ "$EXTRACT" == "c" ]; then
 		git cat-file -p $1 | egrep -i 'secret' | python ${SCRIPT_DIR}/extractor.py --secret
 	fi
 }
@@ -56,6 +73,9 @@ walk_tree() {
 	type=$(git cat-file -t $1)
 	if [ "$type" = "blob" ]; then
 		dump_blob $1
+	elif [ "$type" = "commit" ]; then
+		tree=$(git cat-file -p $1 | grep tree | cut -d " " -f 2)
+		walk_tree $tree
 	else
 		# git cat-file -p $2 | cut -d " " -f 3 | cut -d "	" -f 1
 		subtrees=$(git cat-file -p $1 | cut -d " " -f 3 | cut -d "	" -f 1)
@@ -65,7 +85,7 @@ walk_tree() {
 	fi
 }
 
-while getopts "g:w:f:x:shC" opt; do
+while getopts "g:w:f:x:shCWPr" opt; do
 	case $opt in
 		g)
 			GIT_DIR=$OPTARG
@@ -91,6 +111,16 @@ while getopts "g:w:f:x:shC" opt; do
 			COMMITS=true
 			echo "Printing commit hashes"
 			;;
+		W)
+			OBJECT_HASH=$OPTARG
+			;;
+		P)
+			WALK_PACK=true
+			echo "Walking pack file"
+			;;
+		r)
+			RESUME=true
+			;;
 		h)
 			usage
 			exit
@@ -113,30 +143,59 @@ else
 	exit
 fi
 
+# are we searching for a which commit?
+if [[ $OBJECT_HASH -ne '' ]]; then
+	which_commit $OBJECT_HASH
+	exit
+fi
+
+
+
 # prepare working dir
 if [ -d $WORK ]; then
 	rm $WORK/commit_hashes
-	rm $WORK/tree_hashes
+	if [[ "$RESUME" -ne true ]]; then
+		rm $WORK/tree_hashes
+	fi
 else
 	echo 'Making work directory $WORK'
 	mkdir $WORK
 fi
 
-# get the commit hashes that have $filter
-git log --pretty=tformat:"%H" -- $FILTER > $WORK/commit_hashes
+# if not resuming, get the trees
+if [[ "$RESUME" -ne true ]]; then
 
-# get the trees
-while read line; do
-	if [ -z "$FILTER" ]; then
-		git cat-file -p $line^{tree} | \
-			cut -d " " -f 3 | cut -d "	" -f 1  >> $WORK/tree_hashes
+	if [[ "$WALK_PACK" -eq true ]]; then
+		echo "Walking Pack"
+		echo "This may take awhile...."
+		for f in `ls .git/objects/pack/pack-*.pack`
+		do
+			echo $f
+			git verify-pack -v $f | egrep '(commit|tree|blob)' | cut -d " " -f 1 >> $WORK/commit_hashes
+		done
+		
 	else
-		git cat-file -p $line^{tree} | grep $FILTER | \
-			cut -d " " -f 3 | cut -d "	" -f 1  >> $WORK/tree_hashes
+		# get the commit hashes that have $filter
+		git log --pretty=tformat:"%H" -- $FILTER > $WORK/commit_hashes
 	fi
-	
-done < $WORK/commit_hashes
-	
+
+
+	while read line; do
+		# if [[ "$WALK_PACK" -eq true ]]; then
+		# 	# we already cut it down so just echo the hash over
+		# 	git cat-file -p $line | cut -d " " 
+		# 	echo "$line" >> $WORK/tree_hashes
+		if [ -z "$FILTER" ]; then
+			git cat-file -p $line^{tree} | \
+				cut -d " " -f 3 | cut -d "	" -f 1  >> $WORK/tree_hashes
+		else
+			git cat-file -p $line^{tree} | grep $FILTER | \
+				cut -d " " -f 3 | cut -d "	" -f 1  >> $WORK/tree_hashes
+		fi
+		
+	done < $WORK/commit_hashes
+fi
+
 # iterate through trees looking for blobs
 while read line; do
 	# walk tree with depth 0
